# cmd/server/docs/swagger.yaml
basePath: /
definitions:
  req.EmployeeCreateReq:
    properties:
      id:
        description: ID 字段为自增且后端生成
        example: 1
        type: integer
      idNumber:
        description: |-
          身份证号：必填 + 18位 + 基础正则（支持末尾X/x）
          正则转化： 0x7C -> |  0x2C -> ,
        example: "110105199003078888"
        type: string
      name:
        example: 张三
        type: string
      phone:
        description: 手机号：必填 + 11位 + 大陆手机号正则
        example: "13800000000"
        type: string
      sex:
        description: 性别：限制枚举值
        example: 男
        type: string
      username:
        description: 用户名：必填，唯一
        example: admin
        type: string
    required:
    - idNumber
    - name
    - phone
    - sex
    - username
    type: object
  req.EmployeeLoginReq:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  response.EmployeeLoginResult:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: 管理员
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      userName:
        example: admin
        type: string
    type: object
  response.EmployeePageItemDTO:
    properties:
      createTime:
        example: "2026-04-06 21:30:00"
        type: string
      createUser:
        example: 1
        type: integer
      id:
        example: 1
        type: integer
      idNumber:
        example: 110105********8888
        type: string
      name:
        example: 管理员
        type: string
      phone:
        example: "13800138000"
        type: string
      sex:
        example: 男
        type: string
      status:
        example: 1
        type: integer
      updateTime:
        example: "2026-04-06 21:30:00"
        type: string
      updateUser:
        example: 1
        type: integer
      username:
        example: admin
        type: string
    type: object
  response.PageResult:
    properties:
      records:
        items:
          type: object
        type: array
      total:
        example: 120
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: 编码：1成功，0和其它数字为失败
        example: 1
        type: integer
      data:
        description: 数据
        type: object
      msg:
        description: 错误信息
        example: success
        type: string
    type: object
info:
  contact: {}
  description: 苍穹外卖后端 API 文档
  title: 苍穹外卖
  version: "1.0"
paths:
  /admin/employee:
    post:
      consumes:
      - application/json
      description: 创建员工（需管理员权限）
      parameters:
      - description: 员工信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/req.EmployeeCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
        - BearerAuth: [ ]
      summary: 创建员工
      tags:
      - 员工
  /admin/employee/login:
    post:
      consumes:
      - application/json
      description: 通过用户名和密码登录
      parameters:
      - description: 登录凭证
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/req.EmployeeLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/response.EmployeeLoginResult'
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 员工登录
      tags:
      - 员工
  /admin/employee/page:
    get:
      consumes:
        - application/json
      description: 分页查询员工（需管理员权限）
      parameters:
        - default: '"张三"'
          description: 员工姓名
          in: query
          name: name
          type: string
        - default: 1
          description: 页码
          in: query
          name: page
          required: true
          type: integer
        - default: 10
          description: 每页数量
          in: query
          name: pageSize
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    allOf:
                      - $ref: '#/definitions/response.PageResult'
                      - properties:
                          records:
                            items:
                              $ref: '#/definitions/response.EmployeePageItemDTO'
                            type: array
                        type: object
                type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
        - BearerAuth: [ ]
      summary: 分页查询员工
      tags:
        - 员工
schemes:
  - http
securityDefinitions:
  BearerAuth:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"