# internal/pkg/response/response.go
package response

import (
	"fmt"
	"net/http"
	"sky-take-out-go/internal/pkg/errcode"
	"sky-take-out-go/internal/pkg/i18n"

	"github.com/gin-gonic/gin"
)

// Response 统一响应结构
type Response struct {
	//编码：1成功，0和其它数字为失败
	Code int `json:"code" example:"1"`

	//错误信息
	Message string `json:"msg" example:"success"`

	// 数据
	Data any `json:"data,omitempty" swaggertype:"object"`
}

// Success 成功响应
func Success(c *gin.Context, data any) {
	c.JSON(http.StatusOK, Response{
		Code:    1,
		Message: "success",
		Data:    data,
	})
}

// SuccessWithMsg 成功响应（自定义消息）
func SuccessWithMsg(c *gin.Context, msg string, data any) {
	c.JSON(http.StatusOK, Response{
		Code:    1,
		Message: msg,
		Data:    data,
	})
}

// Error 错误响应（使用 AppError，自动根据语言翻译）
func Error(c *gin.Context, err *errcode.AppError) {
	lang := i18n.GetLang(c)
	msg := i18n.GetMessage(err.Code, lang)
	if msg == "" {
		msg = err.Message
	}

	// 将错误注入 Gin 错误链，让 Logger 中间件能够捕获并记录
	_ = c.Error(fmt.Errorf("[%d] %s", err.Code, msg))

	c.JSON(err.HTTPStatus(), Response{
		Code:    err.Code,
		Message: msg,
	})
}

// ErrorWithMsg 错误响应（自定义）
func ErrorWithMsg(c *gin.Context, httpStatus int, code int, msg string) {
	// 将错误注入 Gin 错误链，让 Logger 中间件能够捕获并记录
	_ = c.Error(fmt.Errorf("[%d] %s", code, msg))

	c.JSON(httpStatus, Response{
		Code:    code,
		Message: msg,
	})
}

// PageResult 分页数据
type PageResult struct {
	Records any   `json:"records" swaggertype:"array,object"`
	Total   int64 `json:"total" example:"120"`
}

// SuccessWithPage 分页成功响应
func SuccessWithPage(c *gin.Context, records any, total int64, page, size int) {
	Success(c, PageResult{
		Records: records,
		Total:   total,
	})
}